bug fixed
nr of lessons was not stored in Database, fixed!

I also fixed a weird issue: the form for LessonBlockForm made text a required field, but I don know why. It happend only for block not for lessons or steps. I implemented a workaround, setting the fields required attribute to False in the form.